+++ quil-py/quil/program/__init__.py
@@ -1,0 +1,5 @@
+"""
+The `quil.program` module contains classes for constructing and representing a 
+Quil program.
+"""
+from quil.program import *