+++ quil-py/src/program/mod.rs
@@ -24,7 +24,10 @@ use rigetti_pyo3::{
 
 use crate::{
     impl_eq, impl_to_quil,
-    instruction::{PyDeclaration, PyGateDefinition, PyInstruction, PyQubit, PyWaveform},
+    instruction::{
+        PyDeclaration, PyGateDefinition, PyInstruction, PyMemoryReference, PyQubit, PyTarget,
+        PyWaveform,
+    },
 };
 
 pub use self::{calibration::PyCalibrationSet, frame::PyFrameSet, memory::PyMemoryRegion};
@@ -234,8 +237,36 @@ impl PyProgram {
             .to_owned())
     }
 
+    pub fn filter_instructions(&self, py: Python, predicate: Py<PyFunction>) -> PyResult<Self> {
+        let filtered = self.as_inner().filter_instructions(|inst| {
+            Python::with_gil(|py| {
+                predicate
+                    .call1(py, (inst.to_python(py).unwrap(),))
+                    .unwrap_or_else(|err| panic!("predicate function returned an error: {err}"))
+                    .extract(py)
+                    .unwrap_or_else(|err| panic!("predicate function must return a bool: {err}"))
+            })
+        });
+        filtered.to_python(py)
+    }
+
     pub fn resolve_placeholders(&mut self) {
         self.as_inner_mut().resolve_placeholders();
+    }
+
+    pub fn wrap_in_loop(
+        &self,
+        loop_count_reference: PyMemoryReference,
+        start_target: PyTarget,
+        end_target: PyTarget,
+        iterations: u32,
+    ) -> Self {
+        PyProgram(self.as_inner().wrap_in_loop(
+            loop_count_reference.into_inner(),
+            start_target.into_inner(),
+            end_target.into_inner(),
+            iterations,
+        ))
     }
 
     // Because we can't bubble up an error from inside the closures, they panic when the given